+++ src/transporters/tcp.js
@@ -109,8 +109,6 @@ class TcpTransporter extends Transporter {
 	 * @memberof TcpTransporter
 	 */
 	connect() {
-		this.logger.warn(kleur.yellow().bold("TCP Transporter is an EXPERIMENTAL transporter. Do NOT use it in production yet!"));
-
 		return Promise.resolve()
 			.then(() => {
 				// Load offline nodes