+++ dev/caller.js
@@ -20,6 +20,13 @@ broker1.createService({
 			this.logger.info(`The '${kleur.cyan().bold(ctx.action.name)}' action is called from '${kleur.yellow().bold(ctx.caller)}' of '${ctx.nodeID}'`);
 			return ctx.call("solar.hello");
 		},
+	},
+
+	events: {
+		async "some.thing"(ctx) {
+			this.logger.info(`The '${kleur.cyan().bold(ctx.event.name)}' event is called from '${kleur.yellow().bold(ctx.caller)}' of '${ctx.nodeID}'`);
+			return ctx.call("planet.welcome");
+		}
 	}
 });
 
@@ -43,7 +50,11 @@ broker2.createService({
 broker2.createService({
 	name: "planet",
 	actions: {
-		hello(ctx) {
+		async hello(ctx) {
+			this.logger.info(`The '${kleur.cyan().bold(ctx.action.name)}' action is called from '${kleur.yellow().bold(ctx.caller)}' of '${ctx.nodeID}'`);
+			await ctx.emit("some.thing");
+		},
+		async welcome(ctx) {
 			this.logger.info(`The '${kleur.cyan().bold(ctx.action.name)}' action is called from '${kleur.yellow().bold(ctx.caller)}' of '${ctx.nodeID}'`);
 		},
 	}
@@ -52,9 +63,9 @@ broker2.createService({
 
 broker1.Promise.all([broker1.start(), broker2.start()])
 	.delay(1000)
-	.then(() => {
+	.then(async () => {
 		broker1.repl();
 
 		broker1.logger.info("");
-		broker1.call("galaxy.hello");
+		await broker1.call("galaxy.hello");
 	});